moved is_inbound_allowed
